custom public location table - updated config_requestblank file and include file for controlcard template
Display a table for public-facing location information.  Transforms archon locations into location codes and then uses those to look up descriptive text defined in config_request file. Also used to generate request links with the proper location code, but it is not necessary to turn on the request links to use this table.